Split out the mongo tester for reuse
