+++ src/us/kbase/test/auth2/lib/storage/mongo/MongoStorageTester.java
@@ -1,0 +1,44 @@
+package us.kbase.test.auth2.lib.storage.mongo;
+
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+
+import com.mongodb.MongoClient;
+import com.mongodb.client.MongoDatabase;
+
+import us.kbase.auth2.lib.storage.mongo.MongoStorage;
+import us.kbase.common.test.controllers.mongo.MongoController;
+import us.kbase.test.auth2.TestCommon;
+
+public class MongoStorageTester {
+
+	static MongoController mongo;
+	static MongoClient mc;
+	static MongoDatabase db;
+	static MongoStorage storage;
+	
+	@BeforeClass
+	public static void beforeClass() throws Exception {
+		TestCommon.stfuLoggers();
+		mongo = new MongoController(TestCommon.getMongoExe().toString(),
+				TestCommon.getTempDir(),
+				TestCommon.useWiredTigerEngine());
+		mc = new MongoClient("localhost:" + mongo.getServerPort());
+		db = mc.getDatabase("test_mongostorage");
+		storage = new MongoStorage(db);
+	}
+	
+	@AfterClass
+	public static void tearDownClass() throws Exception {
+		mc.close();
+		if (mongo != null) {
+			mongo.destroy(TestCommon.isDeleteTempFiles());
+		}
+	}
+	
+	@Before
+	public void clearDB() throws Exception {
+		TestCommon.destroyDB(db);
+	}
+}
